refactor : 클래스 명 변경  schedule 패키지 안으로 이동
