+++ src/main/java/com/kakaotech/team14backend/outer/post/schedule/SchedulePostPopularity.java
@@ -1,21 +1,20 @@
-package com.kakaotech.team14backend.inner.post.usecase;
+package com.kakaotech.team14backend.outer.post.schedule;
 
+import com.kakaotech.team14backend.inner.post.usecase.UpdatePostPopularityUsecase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 
-
 @Component
 @RequiredArgsConstructor
-public class SchedulePostViewCountUsecase {
+public class SchedulePostPopularity {
 
-  private final UpdatePostViewCountUsecase updatePostViewCountUsecase;
+  private final UpdatePostPopularityUsecase updatePostPopularityUsecase;
 
   @Scheduled(initialDelayString = "${schedules.initialDelay}",fixedDelayString = "${schedules.fixedDelay}")
   public void execute() {
-    updatePostViewCountUsecase.execute();
+    updatePostPopularityUsecase.execute();
   }
-
 
 }
